Version 1.16

+++ examples/ReceiveDataInChunks.cpp
@@ -22,7 +22,7 @@ int main(int argc, char **argv) {
 		bool flush = argc > 3;
 		// resolve the stream of interest & make an inlet
 		lsl::stream_info inlet_info = lsl::resolve_stream("name", name).at(0);
-		lsl::stream_inlet inlet(inlet_info, max_buffered);
+		lsl::stream_inlet inlet(inlet_info, (int32_t)max_buffered);
 
 		// Use set_postprocessing to get the timestamps in a common base clock.
 		// Do not use if this application will record timestamps to disk -- it is better to
